+++ crates/target/doc/src-files.js
@@ -1,0 +1,2 @@
+createSrcSidebar('[["flux_composition_graph",["",[],["lib.rs"]]],["flux_text_layout",["",[],["lib.rs"]]]]');
+//{"start":19,"fragment_lengths":[45,40]}