+++ crates/target/doc/trait.impl/core/cmp/trait.Ord.js
@@ -1,0 +1,9 @@
+(function() {
+    const implementors = Object.fromEntries([["flux_composition_graph",[["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/1.97.1/core/cmp/trait.Ord.html\" title=\"trait core::cmp::Ord\">Ord</a> for <a class=\"struct\" href=\"flux_composition_graph/struct.NodeId.html\" title=\"struct flux_composition_graph::NodeId\">NodeId</a>",0],["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/1.97.1/core/cmp/trait.Ord.html\" title=\"trait core::cmp::Ord\">Ord</a> for <a class=\"struct\" href=\"flux_composition_graph/struct.TargetId.html\" title=\"struct flux_composition_graph::TargetId\">TargetId</a>",0]]]]);
+    if (window.register_implementors) {
+        window.register_implementors(implementors);
+    } else {
+        window.pending_implementors = implementors;
+    }
+})()
+//{"start":59,"fragment_lengths":[572]}